requirements_producao.txt
# Dependências para Produção - Sistema de Gestão de Projetos v2.0.0
# Versão refatorada com estrutura modular

# Framework principal
Flask==3.1.1
Werkzeug>=3.1.0

# Banco de dados
supabase==2.3.4
psycopg2-binary==2.9.9

# Autenticação e segurança
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
cryptography==42.0.5

# Sessões e cache
Flask-Session==0.5.0
Flask-Caching==2.1.0

# Rate limiting e segurança
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# Compressão e otimização
Flask-Compress==1.14

# Logging e monitoramento
structlog==24.1.0
sentry-sdk[flask]==2.15.0

# Utilitários
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.4.0

# Desenvolvimento e testes (opcional)
pytest==8.0.2
pytest-flask==1.3.0
black==24.1.1
flake8==7.0.0

# Gunicorn para produção
gunicorn==21.2.0
gevent==23.9.1 